+++ tests/ops/test_aggregation_ops.py
@@ -5,6 +5,8 @@ import pytest
 from trane.ops.aggregation_ops import (
     AvgAggregationOp,
     CountAggregationOp,
+    FirstAggregationOp,
+    LastAggregationOp,
     MajorityAggregationOp,
     MaxAggregationOp,
     MinAggregationOp,
@@ -53,6 +55,20 @@ def test_min_agg_op(df):
     output = op(df)
     assert output == np.min(df["col"])
     assert "the minimum <col> in all related records" in op.generate_description()
+
+
+def test_first_agg_op(df):
+    op = FirstAggregationOp("col")
+    output = op(df)
+    assert output == df["col"].iloc[0]
+    assert "the first <col> in all related records" in op.generate_description()
+
+
+def test_last_agg_op(df):
+    op = LastAggregationOp("col")
+    output = op(df)
+    assert output == df["col"].iloc[-1]
+    assert "the last <col> in all related records" in op.generate_description()
 
 
 @pytest.mark.parametrize(